% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow}
\alias{arrow}
\title{Find upset, downset of a set among a lattice}
\usage{
arrow(x, M, method = NULL)
}
\arguments{
\item{x}{A binary vector encoding an element of a poset}

\item{M}{Binary matrix encoding the lattice from which the supset or downset is found}

\item{method}{Return supremum or infimum}
}
\value{
b Binary vector of sup or inf
}
\description{
Find upset, downset of a set among a lattice
}
\examples{
M <- matrix(c(0,1,1,1,1,1),nrow=2)
x <- c(0,1,0)
arrow(x,M,"up")
}
\author{
Peiyuan Zhu
}
