% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSMprint.R
\name{bcaPrint}
\alias{bcaPrint}
\title{DSMprint is  the new function name of function bcaPrint
#' @param x A list of class bcaspec.}
\usage{
bcaPrint(x, remove = FALSE)
}
\arguments{
\item{remove}{Default = FALSE. Put = TRUE to exclude subsets with zero mass.}
}
\value{
A table of subsets with their associated mass. Subsets are identified by row names.
}
\description{
DSMprint is  the new function name of function bcaPrint
#' @param x A list of class bcaspec.
}
\author{
Claude Boivin
}
