% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcaPrintL.R
\name{bcaPrintL}
\alias{bcaPrintL}
\title{Print summary statistics of large mass functions}
\usage{
bcaPrintL(x, num_cut = 10, num_top_mass = 10)
}
\arguments{
\item{x}{A basic chance assignment (see \code{\link{bca}}).}

\item{num_cut}{= 10 number of groups with equal number of subsets}

\item{num_top_mass}{= 10 number of top masses to be printed}
}
\value{
\itemize{
  \item tables of basic and more comprehensive statistics of subsets
  \item tables of basic and more comprehensive statistics of masses
  \item tables of basic and more comprehensive statistics of masses vs subsets
}
}
\description{
Print summary statistics of large mass functions
}
\examples{
x <- bca(tt = matrix(c(1,1,0,1,1,1), nrow = 2, byrow = TRUE), m = c(0.8, 0.2), cnames = c(1,2,3))
bcaPrintL(x)
}
\author{
Peiyuan Zhu
}
