% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcaRel.R
\name{bcaRel}
\alias{bcaRel}
\title{Representation of a belief function in a product space}
\usage{
bcaRel(tt, spec, infovar, infovarnames = NULL, relnb = NULL)
}
\arguments{
\item{tt}{A (0,1) or logical matrix establishing the relation between two or more variables.}

\item{spec}{A two column matrix. First column contains specification numbers. Second column contains the mass associated with each element of the relation. The same specification number with its associated mass can be repeated (e.g. in a logical implication).}

\item{infovar}{A two column matrix containing variable identification numbers and the number of elements of each variable. The identification numbers must be ordered in increasing number.}

\item{infovarnames}{The names of the variables. if omitted, variables are named v1, v2, etc.}

\item{relnb}{A number given to the relation. Set at 0 if omitted.}
}
\value{
An object of class \code{bcaspec}. This is a list containing the following elements:  \itemize{
\item $con The measure of conflict. Set at 0 by default.
\item $tt The resulting table of focal elements alone. Rownames of the matrix of focal elements are created from the column names of the elements of the product frame.
\item $spec The resulting two column matrix of specification numbers with associated masses.
\item $infovar The two column matrix given in the input data.
\item infovaluenames A list of the names of the variables with the value name of each element.
\item $inforel. A two column matrix containing variable numbers and the depth of the relation.
}
}
\description{
When a relation between two or more variables is established, a product space representation of this relation can be obtained. The relation is described by a matrix input table of (0,1), where the different spaces are side by side, as a truth table representation.
}
\examples{
## A logical implication table.
## A typical relation between two variables is the logical implication (a -> b) or its conditional probability equivalent (b | a). let us suppose a stands for Rain (yes, no) and b stands for Road Works (yes, no). From our experience, we are 75 \% sure that there will be Road Works if no rain.
## The truth table
 ttrwf= matrix(c(0,1,1,0,1,0,1,0,1,0,0,1,1,1,1,1),nrow=4, byrow = TRUE, dimnames =list(NULL, c("rWdy", "rWdn", "Ry", "Rn")) )
 ## The mass distribution
 specrw = matrix(c(1,1,1,2,0.75,0.75,0.75,0.25), ncol = 2, dimnames = list(NULL, c("specnb", "mass"))) 
 ## variables numbers and sizes
 inforw =matrix(c(4,5,2,2), ncol = 2, dimnames = list(NULL, c("varnb", "size")) )
bcaRel(tt = ttrwf, spec = specrw, infovar = inforw, infovarnames = c("RdWorks", "Rain"), relnb = 6)
 
}
\author{
Claude Boivin, Stat.ASSQ
}
