% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/belplauEval.R
\name{belplauEval}
\alias{belplauEval}
\title{Evaluate type I, II errors}
\usage{
belplauEval(
  bel_plau,
  true_order,
  var = "rplau",
  err = "first kind",
  is_belplau = TRUE
)
}
\arguments{
\item{bel_plau}{belplau object or a vector whose ordering is compared}

\item{true_order}{a vector representing the true ordering}

\item{var}{="rplau" variable name of the belplau to be used as ordering}

\item{err}{="first kind" type of error to be evaluated}

\item{is_belplau}{=TRUE whether bel_plau is a belplau object or just a numerical vector}
}
\value{
first kind, second kind, or total error
}
\description{
Calculate first kind, second kind, total error by comparing two vectors. One vector represents the truth and the other represents a numerical quantity of importance. 
Error of the first kind: out of all the comparisons between relevant and irrelevant elements, what proportion of errors are due to indicating an irrelevant element as more important than a relevant element
Error of the second kind: out of all the comparisons between relevant and irrelevant elements, what proportion of errors are due to indicating an relevant element as less important than an irrelevant element
Total error: the sum of error of the first kind and the error of the second kind
}
\examples{
x <- bca(tt = matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, 
byrow = TRUE), m = c(0.2,0.5, 0.3), 
cnames = c("a", "b", "c"), varnames = "x", idvar = 1)
belplau(x)
y <- bca(tt = matrix(c(1,0,0,1,1,1),nrow = 2, 
byrow = TRUE), m = c(0.6, 0.4),  
cnames = c("a", "b", "c"),  varnames = "y", idvar = 1)
xy <- nzdsr(dsrwon(x,y))
z<-belplau(xy,h=ttmatrixPartition(xy$infovar[2],xy$infovar[2]))
belplauEval(z,c(0,1,0))
}
\author{
Peiyuan Zhu
}
