% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/belplauPlot.R
\name{belplauPlot}
\alias{belplauPlot}
\title{Plot belplau matrix}
\usage{
belplauPlot(belplau_mat, xlab, color, y = "rplau", x = "index")
}
\arguments{
\item{belplau_mat:}{belplau matrix e.g. belplau(bpa)}

\item{xlab:}{x-axis labels e.g. c("1:34","35:68","69:101")}

\item{color:}{color of xlab e.g. c(0,1,0)}

\item{y="rplau":}{column name of belplau matrix}
}
\value{
a plot of belplau matrix
}
\description{
Plot belplau matrix
}
\examples{
bpa <- bca(tt = matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, 
byrow = TRUE), m = c(0.2,0.5, 0.3), 
cnames = c("a", "b", "c"), varnames = "x", idvar = 1)
bel_plau <- belplau(bpa)
belplauPlot(bel_plau)
}
\author{
Peiyuan Zhu
}
