% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{bound}
\alias{bound}
\title{Find inf, sup of a lattice}
\usage{
bound(x, method = NULL)
}
\arguments{
\item{x}{Binary matrix encoding the lattice from which the supremum or infimum is found}

\item{method}{Return supremum or infimum}
}
\value{
b Binary vector of sup or inf
}
\description{
Find inf, sup of a lattice
}
\examples{
M <- matrix(c(0,1,1,1,1,1),nrow=2)
x <- t(as.matrix(c(0,1,0)))
bound(M,"inf")
}
\author{
Peiyuan Zhu
}
