% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_outputs.R
\name{buildBaseQuota}
\alias{buildBaseQuota}
\title{Build base quota}
\usage{
buildBaseQuota(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Reads the assessment outputs and builds a GUI
plotting interface to display the important
results of the base assessment.
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
