% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_outputs.R
\name{buildKobeMatrix}
\alias{buildKobeMatrix}
\title{Build Kobe Matrix}
\usage{
buildKobeMatrix(input, output, session, iteration, MSYoutput = NULL)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}

\item{iteration}{current iteration of loop.}

\item{MSYoutput}{Null in first iteration where MSY
calculated. For later iterations this is the output
of the first iteration.}
}
\description{
Iterates over a numer of assessments and fixed catch
rates to calculate a probability matrix of overfishing
and overfished likelihoods.
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
