% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_interface.R
\name{buildManagementTabPanels}
\alias{buildManagementTabPanels}
\title{Build tab panels}
\usage{
buildManagementTabPanels(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Build the tab panels for each group
}
\author{
Nathan Vaughan
}
\keyword{building}
\keyword{interface}
