% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_interface.R
\name{buildTabElements}
\alias{buildTabElements}
\title{Build tab elements}
\usage{
buildTabElements(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Build the UI elements for each tab of the shiny interface
}
\author{
Nathan Vaughan
}
\keyword{building}
\keyword{interface}
