% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTree.R
\name{buildTree}
\alias{buildTree}
\title{Build a tree of...}
\usage{
buildTree(tt, qq, indices = NULL)
}
\arguments{
\item{tt}{A (0,1)-matrix or a boolean matrix. The number of columns must match the number of elements (values) of the Frame of Discernment (FoD).}

\item{q}{Commonality values of the \code{tt} matrix.}
}
\value{
tree The tree structure.
}
\description{
Build a tree of...
}
\details{
A tree structure of commonalities is built from the binary matrix \code{tt}. This tree structure can then be used with large frames of discernment to optimize the combination of support functions.
}
\examples{
# Example from figure 12 of the cited reference.
 x <- matrix(c(1,0,0,
 0,0,1,
 0,1,1,
 1,1,1), nrow = 4, byrow = TRUE, dimnames = list(NULL,c("a","b","c")))
 rownames(x) <- nameRows(x)
 m <- c(0.1,0.2,0.3,0.4)
 q <- commonality(x,m,"ezt-m")
 q_tree <- buildTree(x,q)
}
\references{
Chaveroche, Franck Davoine, Véronique Cherfaoui. Eﬃcient Möbius Transformations and their applications to Dempster-Shafer Theory: Clarification and implementation. ArXiv preprint arXiv:2107.07359.
}
\author{
Peiyuan Zhu
}
