% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{closure}
\alias{closure}
\title{Augment list of binary vectors with closure elements}
\arguments{
\item{ttx}{A binary matrix}

\item{computeJoin}{= true: to compute join closure. Default = TRUE}

\item{display_progress}{= true: to compute join closure. Default = FALSE}
}
\value{
A list of binary vectors including the closure elements
}
\description{
Augment list of binary vectors with closure elements
}
