% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{closureSparse}
\alias{closureSparse}
\title{Augment a sparse binary matrixs with closure elements}
\arguments{
\item{ttx}{A sparse binary matrix}

\item{computeJoin}{= true: to compute join closure. Default = TRUE}

\item{display_progress}{= true: to compute join closure. Default = FALSE}
}
\value{
A sparse binary matrix including the closure elements
}
\description{
Augment a sparse binary matrixs with closure elements
}
