% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{commSparse}
\alias{commSparse}
\title{Comptue commonality values of a group of simple support functions on its closure elements}
\arguments{
\item{x}{The non-trivial support of the mass function}

\item{x_c}{A sparse binary matrix of closure elements}

\item{a}{Mass assigned to the non-trivial support}

\item{display_progress}{= true: to show progress bar. Default = FALSE}
}
\value{
A sparse binary matrix including the closure elements
}
\description{
Comptue commonality values of a group of simple support functions on its closure elements
}
