% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{composition}
\alias{composition}
\title{Compose two relation}
\usage{
composition(R, S)
}
\arguments{
\item{R}{first relation from set A to set B to be combined, encoded by a binary matrix}

\item{S}{second relation from set B to set C to be combined, encoded by a binary matrix}
}
\value{
The Schöder-composed relation \deqn{J = \max_{j\in B}(\min(R_{i,j}, S_{j,k}))} indicating whether there exists an intermediate element connecting the two elements from A and B
}
\description{
Calculate Schöder's composition of two relations
}
\examples{
R1 <- matrix(c(1,1,0,0,1,1,0,0,1), nrow=3, byrow = TRUE)
composition(R1,R1)
}
\author{
Peiyuan Zhu
}
