% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNode.R
\name{createNode}
\alias{createNode}
\title{Change depth to appropriate calculation
- depth equals the greatest index corresponding to 1 in the bit vector x}
\usage{
createNode(x, q)
}
\description{
Change depth to appropriate calculation
- depth equals the greatest index corresponding to 1 in the bit vector x
}
