% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoSSnames.R
\name{DoSSnames}
\alias{DoSSnames}
\title{Construct subsets names from column names of a tt matrix}
\usage{
DoSSnames(tt)
}
\arguments{
\item{tt}{A description matrix with column names}
}
\value{
subsets_names A list of names.
}
\description{
Construct subsets names from column names of a tt matrix
}
\examples{
y1 <- bca(tt = matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, 
byrow = TRUE), m = c(0.2,0.5, 0.3), 
cnames = c("a", "b", "c"),  
varnames = "x", idvar = 1) 
DoSSnames(y1$tt)
}
\author{
Claude Boivin
}
