% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSA.R
\name{DSA}
\alias{DSA}
\alias{belplau}
\title{Calculate the degrees of unambiguous support (p), unambiguous contradiction (q), and ambiguity (r) of a list of hypothesis based on a Dempster-Shafer Model (DSM).}
\usage{
DSA(x, remove = FALSE, h = NULL)

belplau(x, remove = FALSE, h = NULL)
}
\arguments{
\item{x}{A DSM (see \code{\link{DSM}}).}

\item{remove}{= TRUE: Exclude subsets with zero mass.}

\item{h}{= NULL: Hypothesis to be tested. Description matrix in the same format than \code{x$tt}}
}
\value{
A matrix of \code{M} rows by 3 columns is returned, where \code{M} is the number of focal sets: \itemize{
 \item Column 1: the degree of unambiguous support (Belief) \code{p};
 \item Column 2: the degree of unambiguous contradiction (Disbelief) \code{q};
 \item Column 3: the degree of ambiguity (Epistemic uncertainty) \code{r};
 \item Column 4: the degree of non-contradiction (Plausibility) \code{plau};
 \item Column 5: the Plausibility ratio \code{rplau}.
   }
}
\description{
Degrees of unambiguous support \code{p} and unambiguous contradiction \code{q} of the focal sets of a DSM are computed. The ratio of the plausibility of a focal set against the plausibility of its contrary is also computed. Subsets with zero mass can be excluded from the calculations.\cr
}
\details{
The degree of unambiguous support (belief) \code{p} is defined by: \cr
\deqn{p(A) = Sum((m(B); B \subseteq A))}{p(A) = Sum((m(B); B <= A))} for every subset B of A.\cr
The degree of non-contradiction (plausibility) \code{plau} is defined by: \cr
\deqn{plau(A) = Sum[(m(B); B \cap A \neq \emptyset]}{pl(A) = Sum[(m(B); (B & A) not empty]} for every subset \code{B} of the state-space model. \cr
The plausibility ratio of a focal set \code{A} versus its contrary \code{not A} is defined by:  \eqn{(1-q(A))/(1-p(A))}.
}
\examples{
x <- DSM(tt = matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, 
byrow = TRUE), m = c(0.2,0.5, 0.3), 
cnames = c("a", "b", "c"), varnames = "x", idvar = 1)
DSA(x)
y <- DSM(tt = matrix(c(1,0,0,1,1,1),nrow = 2, 
byrow = TRUE), m = c(0.6, 0.4),  
cnames = c("a", "b", "c"),  varnames = "y", idvar = 1)
xy <- normalize(DSC(x,y))
DSA(xy)
print("compare all elementary events")
xy1 <- addToDSM(x = xy, tt = matrix(c(0,1,0,0,0,1), nrow = 2, byrow = TRUE))
DSA(xy1) 
DSA(xy1, remove = TRUE) 
DSA(xy1, h = matrix(c(1,0,0,0,1,1), nrow = 2, byrow = TRUE))

}
\references{
\itemize{
  \item Shafer, G., (1976). A Mathematical Theory of Evidence. Princeton University Press, Princeton, New Jersey, p. 39-43.
  \item Williams, P., (1990). An interpretation of Shenoy and Shafer's axioms for local computation. International Journal of Approximate Reasoning 4, pp. 225-232.
  \item Dempster, A., (2008). The Dempster–Shafer calculus for statisticians. International Journal of approximate reasoning, 48(2), 365-377.
}
}
\author{
Claude Boivin, Peiyuan Zhu
}
