% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAh.R
\name{DSAh}
\alias{DSAh}
\alias{belplauH}
\title{Calculate degree of unambiguous support (p), unambiguous contradiction (q), ambiguity (r), plausibility, and plausibility ratio}
\usage{
DSAh(MACC, W2, h)

belplauH(MACC, W2, h)
}
\arguments{
\item{MACC}{Vector of masses e.g. x$spec[,2]}

\item{W2}{Description matrix e.g. x$tt}

\item{h}{Hypotheses to be tested, same format as x$tt}
}
\value{
A matrix of \code{M} rows by 5 columns is returned, where \code{M} is the number of hypothesis tested: \itemize{
 \item Column 1: the degree of unambiguous support \code{p};
 \item Column 2: the degree of unambiguous contradiction \code{q};
 \item Column 3: the degree of ambiguity \code{r};
 \item Column 4: the degree of Plausibility \code{plau};
 \item Column 5: the Plausibility ratio \code{rplau}.
   }
}
\description{
Calculate degree of unambiguous support (p), unambiguous contradiction (q), ambiguity (r), plausibility, and plausibility ratio
}
\examples{
x <- DSM(tt = matrix(c(1,1,0,1,1,1), nrow = 2, byrow = TRUE), m = c(0.8, 0.2), cnames = c(1,2,3))
DSAh(MACC = x$spec[,2], W2 = x$tt, h = x$tt)
hyp <- matrix(c(0,1,0,0,1,1), nrow = 2, byrow = TRUE)
rownames(hyp) <- nameRows(hyp)
DSAh(MACC = x$spec[,2], W2 = x$tt, h = hyp)
}
\author{
Peiyuan Zhu
}
