% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC.R
\name{DSC}
\alias{DSC}
\alias{dsrwon}
\alias{The}
\alias{calculations}
\alias{make}
\alias{use}
\alias{of}
\alias{multiple}
\alias{cores}
\alias{available.}
\title{Combination of two Dempster-Shafer models}
\usage{
DSC(
  x,
  y,
  mcores = "no",
  use_ssnames = FALSE,
  use_sparse = "no",
  varnames = NULL,
  relnb = NULL,
  infovarnames
)

dsrwon(
  x,
  y,
  mcores = "no",
  use_ssnames = FALSE,
  use_sparse = "no",
  varnames = NULL,
  relnb = NULL,
  infovarnames
)
}
\arguments{
\item{x}{A Dempster-Shafer model (see \code{\link{DSM}}).}

\item{y}{A Dempster-Shafer model (see \code{\link{DSM}}).}

\item{mcores}{Make use of multiple cores ("yes") or not ("no"). Default = "no".}

\item{use_ssnames}{= TRUE to use ssnames instead of tt matrix to do the intersections. Default = FALSE}

\item{use_sparse}{Make use of sparse matrices ("yes") or not ("no"). Default = "no".}

\item{varnames}{A character string to name the resulting variable. named "z" if omitted.}

\item{relnb}{Identification number of the jointDSM. Can be omitted.}

\item{infovarnames}{Deprecated. Old name for \code{varnames}.}
}
\value{
A Dempster-Shafer model with these two components added: \itemize{
  \item I12 Intersection table of subsets.
  \item Sort_order Sort order of subsets.
  }
}
\description{
The unnormalized Dempster's rule is used to combine two mass functions \code{mx} and \code{my} defined on the same state-space model (SSM) and described by their respective basic mass assignments \code{x} and \code{y}. Dempster's rule of combination is applied. The normalization is not done, leaving the choice to the user to normalize the results or not (for the normalization operation, see function \code{\link{normalize}}).
}
\details{
The two DSM's \code{x} and \code{y} must be defined on the same SSM for the combination to take place. The jointDSM number of the x input is given to the output result.
}
\examples{
y1 <- DSM(tt = matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, 
byrow = TRUE), m = c(0.2,0.5, 0.3), 
cnames = c("a", "b", "c"),  
varnames = "x", idvar = 1) 
y2 <- DSM(tt = matrix(c(1,0,0,1,1,1),nrow = 2, 
byrow = TRUE), m = c(0.6, 0.4),  
cnames = c("a", "b", "c"),  
varnames = "x", idvar = 1)
DSC(y1,y2)
# Sparse matrices
y1s <- y1
y2s <- y2
y1s$tt <- methods::as(y1$tt, "RsparseMatrix")
y2s$tt <- methods::as(y2$tt, "RsparseMatrix")
y1y2s <- DSC(y1s, y2s, use_ssnames = TRUE)
vacuous <- DSM(matrix(c(1,1,1), nrow = 1), m = 1, cnames = c("a","b","c"))
DSC(vacuous, vacuous)
}
\references{
\itemize{
  \item Shafer, G., (1976). A Mathematical Theory of Evidence. Princeton University Press, Princeton, New Jersey, p. 38: Basic probability assignment.
  \item Guan, J. W. and Bell, D. A., (1991). Evidence Theory and its Applications. Elsevier Science Publishing company inc., New York, N.Y., p. 29: Mass functions and belief functions 
  \item Dempster, A., (2008). The Dempster–Shafer calculus for statisticians. International Journal of approximate reasoning, 48(2), 365-377.
}
}
\author{
Claude Boivin, Peiyuan Zhu
}
