\name{dst-package}
\alias{dst-package}
\alias{dst}
\docType{package}
\title{
Using Dempster-Shafer Theory for the combination of evidence
}
\description{
An R package for the use of Dempster-Shafer Theory for combination of evidence.
}
\details{
\tabular{ll}{
Package: \tab dst\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2018-01-01\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}
The following operations can be done:
\describe{
  \item{Basic chance assignment}{Statement of a chance
assignments over subsets of the frame of discernment.}
  \item{Dempster's rule}{Combination of two belief functions and normalization}
  \item{belief/plausibility calculations}{Calculation of the measures of belief and plausibility on a belief function}
  \item{Plausibility ratio}{Calculation of the plausibility transformation and the ratio of plausibility of all the elements of a belief function}
  \item{Measure of conflict}{Calculation of the measure of conflict on a serie of beliefs functions combined by Dempster's rule}
}
}
\author{
Claude Boivin, Stat.ASSQ <webapp.cb@gmail.com>

Maintainer: Claude Boivin <webapp.cb@gmail.com>

}
\references{
Shafer, G., (1976). A Mathematical Theory of Evidence. Princeton University Press, Princeton, New Jersey. 296 p.

Cobb, B. R. and Shenoy, P.P. (2006). On the plausibility transformation method for translating belief function models to probability models. Journal of Approximate Reasoning, 41(3), April 2006, 314--330.
}
\keyword{ package }