% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst.R
\docType{package}
\name{dst}
\alias{dst}
\title{Manipulation and combination of belief functions}
\description{
dst allows you to make basic probability assignments on subsets of a set of possibilities (events) and combine these events with Dempster's rule of combination.
}
\details{
The main operations that can be done are: \itemize{
\item definition of a basic chance assignment (bca) distribution on a variable
 \item combination of two bca's defined on the same variable
 \item definition of a bca which establish a relation between two or more variables
 \item extension of a bca
 \item marginalization of a bca }
}
\author{
Claude Boivin, Stat.ASSQ <webapp.cb@gmail.com>
}
