% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment.R
\name{duplicateFleets}
\alias{duplicateFleets}
\title{Duplicate fleets}
\usage{
duplicateFleets(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
This function allows projection of fishery
scenarios where fishing allocation and selectivity
do not match the historical record. It achieves
this goal by replicating all fishing fleets with
"dummy" fleets and setting their historic catch to
E10-7 times the true fleets catch. Allocation is
then shifted to these fleets and their selectivity
patterns can be freely modified.
}
\author{
Nathan Vaughan
}
\keyword{assessment}
\keyword{projection}
