% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extSSM.R
\name{extSSM}
\alias{extSSM}
\alias{extFrame}
\title{Extension of the state-space model of a variable}
\usage{
extSSM(x, use_ssnames = FALSE, lab = NULL)

extFrame(x, use_ssnames = FALSE, lab = NULL)
}
\arguments{
\item{x}{An object of DSM class, i.e. a basic chance assignment defined on one SSM}

\item{use_ssnames}{Default= FALSE. Put TRUE if use of subset names is wanted.}

\item{lab}{A character vector containing the names of the elements to add to the SSM.}
}
\value{
zxtnd The DSM with its SSM extended
}
\description{
This function works on a Dempster-Shafer model (DSM) \code{x} defined on a single state-sapce model (SSM). It allows the addition of new values to the SSM.
}
\examples{
s1_e1 <- DSM(tt = matrix(c(1,0,1,1),nrow = 2, byrow = TRUE), 
m = c(0.6,0.4), cnames = c("S1","S2"), varnames = "v1", idvar = 1) 
s13_names <- extSSM(s1_e1, lab = "S3", use_ssnames =TRUE)
s13 <- extSSM(s1_e1, lab = "S3")
}
\author{
Claude Boivin
}
