% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extSSR.R
\name{extSSR}
\alias{extSSR}
\alias{extFrame}
\title{Extension of a State Space Representation (SSR)}
\usage{
extSSR(x, use_ssnames = FALSE, lab = NULL)

extFrame(x, use_ssnames = FALSE, lab = NULL)
}
\arguments{
\item{x}{An object of DSM class, i.e. a basic chance assignment defined on one SSR.}

\item{use_ssnames}{Default= FALSE. Put TRUE if use of subset names is wanted.}

\item{lab}{A character vector containing the names of the elements to add to the SSR.}
}
\value{
zxtnd The DSM within its SSR extended.
}
\description{
This function works on a marginal DSM \code{x} defined on a single variable. It allows the addition of new values to the SSR.
}
\examples{
s1_e1 <- DSM(tt = matrix(c(1,0,1,1),nrow = 2, byrow = TRUE), 
m = c(0.6,0.4), cnames = c("S1","S2"), varnames = "v1", idvar = 1) 
s13_names <- extSSR(s1_e1, lab = "S3", use_ssnames =TRUE)
s13 <- extSSR(s1_e1, lab = "S3")
s13$valuenames
}
\author{
Claude Boivin
}
