% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment.R
\name{FitParabola}
\alias{FitParabola}
\title{Fit Parabola}
\usage{
FitParabola(x, y)
}
\arguments{
\item{x}{a vector of three spawning stock biomass ratio values.}

\item{y}{a vector of three total yield values.}
}
\description{
Estimates a parabolic fit through 3 paired estimates of
spawning stock biomass ratio and total sustainable yield in
order to predict the spawning stock biomass ratio that will
correspond to maximum sustainable yield.
}
\author{
Nathan Vaughan
}
\keyword{fitting}
\keyword{model}
