% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment.R
\name{getMangementLimits}
\alias{getMangementLimits}
\title{Read management input}
\usage{
getMangementLimits(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Reads the user input management values from
the shiny GUI for modification of the
assessment projection assumptions.
}
\author{
Nathan Vaughan
}
\keyword{assessment}
\keyword{projection}
