% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeQQ.R
\name{imputeQQ}
\alias{imputeQQ}
\title{Impute commonality values based on a closure matrix}
\usage{
imputeQQ(tty, tt1, tt2, q1, q2, tree_type = NULL)
}
\arguments{
\item{tty}{A closure matrix}

\item{tt1}{A q1 support matrix}

\item{tt2}{A q2 support matrix}

\item{q1}{A named vector of commonality values}

\item{q2}{A named vector of commonality values}

\item{tree_type}{tree_type to use M trees ("multiple") or 1 tree ("single"). Default = NULL}
}
\value{
x a list with two elements \itemize{
 \item q1 new commonality vector
 \item q2 new commonality vector
}
}
\description{
Impute commonality values based on a closure matrix
}
\details{
Impute commonality values based on a closure matrix
}
\examples{
tt1 <- matrix(c(1,1,0,1,1,1), byrow=TRUE, nrow=2, dimnames=list(NULL,c("a","b","c")))
tt2 <- matrix(c(0,1,1,1,1,1), byrow=TRUE, nrow=2, dimnames=list(NULL,c("a","b","c")))
tty <- matrix(c(0,1,1,0,1,1,1,1,1,0,1,0), byrow=TRUE, nrow=4, dimnames=list(NULL,c("a","b","c")))
q1 <- c(1,0.9)
q2 <- c(1,0.8)
names(q1) <- nameRows(tt1)
names(q2) <- nameRows(tt2)
x <- imputeQQ(tty,tt1,tt2,q1,q2)
x$q1
x$q2
}
\author{
Peiyuan Zhu
}
