% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_interface.R
\name{initImplementManagement}
\alias{initImplementManagement}
\title{Implement management}
\usage{
initImplementManagement()
}
\description{
Initializes the management settings shiny tab
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
