% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_interface.R
\name{initKobeChoices}
\alias{initKobeChoices}
\title{Kobe Choices}
\usage{
initKobeChoices(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Initializes the kobe matrix assessment selections
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
