% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_interface.R
\name{initQuotaForecast}
\alias{initQuotaForecast}
\title{forecast}
\usage{
initQuotaForecast()
}
\description{
Initializes the assessment results shiny tab
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
