% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_interface.R
\name{initSelectAssessment}
\alias{initSelectAssessment}
\title{Select assessment}
\usage{
initSelectAssessment()
}
\description{
Initializes the select assessment shiny tab
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
