% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersBySSName.R
\name{intersBySSName}
\alias{intersBySSName}
\title{Intersect two vectors of ssnames}
\usage{
intersBySSName(zx, yz)
}
\arguments{
\item{zx}{a vector of ssnames from one bca}

\item{yz}{a vector of ssnames from another bca}
}
\value{
ssnames in the intersection of the two bcas
}
\description{
Intersect two vectors of ssnames
}
\examples{
y1 <- bca(tt = matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, 
byrow = TRUE), m = c(0.2,0.5, 0.3), 
cnames = c("a", "b", "c"),  
varnames = "x", idvar = 1) 
y2 <- bca(tt = matrix(c(1,0,0,1,1,1),nrow = 2, 
byrow = TRUE), m = c(0.6, 0.4),  
cnames = c("a", "b", "c"),  
varnames = "x", idvar = 1)
intersBySSName(y1$ssnames[[1]], y2$ssnames[[2]])
}
\author{
Peiyuan Zhu
}
