% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iotaSparse}
\alias{iotaSparse}
\title{Comptue iota elements of a sparse binary matrix of closure elements}
\arguments{
\item{tt}{A sparse matrix of closure elements}

\item{display_progress}{= true: to show progress bar. Default = FALSE}
}
\value{
A sparse binary matrix of the iota elements
}
\description{
Comptue iota elements of a sparse binary matrix of closure elements
}
