% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dst.R
\name{load_dst}
\alias{load_dst}
\title{load DST}
\usage{
load_dst(input, output, session, app_objects)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}

\item{app_objects}{a list of global parameters maintained in the app}
}
\description{
Loads all the base observers and interface
of the shiny dst.
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
