% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsum.R
\name{logsum}
\alias{logsum}
\title{Adding small probabilities}
\usage{
logsum(l1, l2)
}
\arguments{
\item{l1}{log probabilities}

\item{l2}{log probabilities}
}
\value{
sum of probabilities exp(l1)+exp(l2)
}
\description{
Adding small probabilities
}
\examples{
# sum of two 1e-5
exp(logsum(log(1e-5),log(1e-5)))
}
\author{
Peiyuan Zhu
}
