% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFromQQ.R
\name{mFromQQ}
\alias{mFromQQ}
\title{Construct a mass vector from qq function and tt matrix of focal elements recursively.}
\usage{
mFromQQ(
  qq,
  n = NULL,
  cnames = NULL,
  method = NULL,
  sparse = "no",
  tt = NULL,
  use_pb = FALSE
)
}
\arguments{
\item{qq}{Commonality function}

\item{n}{Frame dimension}

\item{cnames}{A character vector containing the names of the elements of the frame of discernment}

\item{method}{= NULL: Use Fast Mobius Transform ("fmt") or Efficient Mobius Transform ("emt") or Efficient Mobius Transform on a meet-closed subset ("emt-m")}

\item{tt}{A binary matrix.}

\item{use_pb}{Whether to print progress bar.}

\item{sprase}{= c("yes","no") whether to use sparse matrix. Default = "no".}
}
\value{
m A corresponding mass vector
}
\description{
Construct a mass vector from qq function and tt matrix of focal elements recursively.
}
\examples{
tt<- t(matrix(c(1,0,1,1),ncol = 2))
m<- c(.9,.1)
cnames <- c("yes","no")
x<- bca(tt, m, cnames=cnames, method = "fzt")
mFromQQ(x$qq, 2, method = "fmt", cnames = cnames)
}
\author{
Peiyuan Zhu
}
