% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFromQQRecursive.R
\name{mFromQQRecursive}
\alias{mFromQQRecursive}
\title{Construct a mass vector from qq function and ttmatrix of focal elements recursively.}
\usage{
mFromQQRecursive(qq, n, method = NULL, tt = NULL, cnames = NULL)
}
\arguments{
\item{qq}{Commonality function}

\item{n}{Frame dimension}

\item{method}{= NULL: Use Fast Mobius Transform ("fmt") or Efficient Mobius Transform ("emt") or Efficient Mobius Transform on a meet-closed subset ("emt-m")}

\item{tt}{= NULL: Subsets whose masses are evaluated}

\item{cnames}{= NULL: A character vector containing the names of the elements of the frame of discernment}
}
\value{
m A corresponding mass vector
}
\description{
Construct a mass vector from qq function and ttmatrix of focal elements recursively.
}
\examples{
tt<- t(matrix(c(1,0,1,1),ncol = 2))
m<- c(.9,.1)
cnames <- c("yes","no")
x<- bca(tt, m, cnames=cnames)
mFromQQRecursive(x$qq, 2, method = "fmt", cnames = cnames)
}
\author{
Peiyuan Zhu
}
