% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiusInvHQQ.R
\name{mobiusInvHQQ}
\alias{mobiusInvHQQ}
\title{Mobius inversion of commonality function}
\usage{
mobiusInvHQQ(qq, h)
}
\arguments{
\item{qq}{Commonality function}

\item{h}{Hypothesis to be evaluated}
}
\value{
m Mass of the hypothesis
}
\description{
Mobius inversion of commonality function
}
\examples{
x <- bca(tt = matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, byrow = TRUE),
m = c(0.2,0.5, 0.3), cnames = c("a", "b", "c"), varnames = "x", idvar = 1)
qq <- commonality(x$tt,x$spec[,2])
mobiusInvHQQ(qq, matrix(c(0,1,0,1,1,0), nrow = 2, byrow = TRUE))
}
\author{
Peiyuan Zhu
}
