% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameCols.R
\name{nameCols}
\alias{nameCols}
\title{Naming the columns of the \code{tt} matrix}
\usage{
nameCols(valuenames, size)
}
\arguments{
\item{valuenames}{A list of the names of the variables with the name of the elements of their frame of discernment.}

\item{size}{A vector of the size of the variables.}
}
\value{
A character vector of length equal to the sum of the sizes of the variables.
}
\description{
This utility function makes use of the valuenames  and size parameters  of a set of variables to assign values names to the columns of a \code{tt} matrix.
}
\examples{
infoval <- list(A = c("a1", "a2"), B = c("b1", "b2", "b3"))
sizes <- c(2,3)
 print(nameCols(valuenames = infoval, size = sizes) )
}
\author{
Claude Boivin
}
