% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameCols_prod.R
\name{nameCols_prod}
\alias{nameCols_prod}
\title{Naming the columns of the \code{tt} matrix of a product space}
\usage{
nameCols_prod(valuenames, size)
}
\arguments{
\item{valuenames}{A list of the names of the variables with the name of the elements of their frame of discernment.}

\item{size}{A vector of the size of the variables.}
}
\value{
A character vector of length equal to the product of the sizes of the variables.
}
\description{
This utility function makes use of the valuenames  and size parameters  of a set of variables to assign values names to the columns of the \code{tt} matrix of their product space.
}
\examples{
infoval <- list(A = c("a1", "a2"), B = c("b1", "b2", "b3"))
sizes <- c(2,3)
 print(nameCols_prod(valuenames = infoval, size = sizes) )
}
\author{
Claude Boivin, Stat.ASSQ
}
