% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ui.R
\name{newBSModal}
\alias{newBSModal}
\title{new modal}
\usage{
newBSModal(location, id, placement = "beforeEnd", label,
  hidden = FALSE, immediate = FALSE, ui = NULL)
}
\arguments{
\item{location}{The existing object that the new shiny object will
be placed relative to.}

\item{id}{the id of the new shiny object.}

\item{placement}{Object placement relative to location object.}

\item{label}{the displayed label.}

\item{hidden}{Should the object be hidded when created. Default = FALSE.}

\item{immediate}{Should the object be created immediately. Default = FALSE.}

\item{ui}{the ui to be included in the modal}
}
\description{
Inserts a new shiny BS modal object from the
server code. Wraps the insert UI function.
}
\author{
Nathan Vaughan
}
\keyword{UI}
\keyword{insertion}
