% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ui.R
\name{newCheckbox}
\alias{newCheckbox}
\title{new checkbox}
\usage{
newCheckbox(location, id, label = "", width = NULL, value = FALSE,
  placement = "beforeEnd", hidden = FALSE, immediate = FALSE)
}
\arguments{
\item{location}{The existing object that the new shiny object will
be placed relative to.}

\item{id}{the id of the new shiny object.}

\item{label}{the displayed label.}

\item{width}{width of checkbox.}

\item{value}{true or false.}

\item{placement}{Object placement relative to location object.}

\item{hidden}{Should the object be hidded when created. Default = FALSE.}

\item{immediate}{Should the object be created immediately. Default = FALSE.}
}
\description{
Inserts a new shiny checkbox object from the
server code. Wraps the insert UI function.
}
\author{
Nathan Vaughan
}
\keyword{UI}
\keyword{insertion}
