% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ui.R
\name{newCheckBoxGroup}
\alias{newCheckBoxGroup}
\title{new checkbox group}
\usage{
newCheckBoxGroup(location, id, label = NULL, placement = "beforeEnd",
  choices = NULL, selected = NULL, choiceNames = NULL,
  choiceValues = NULL, inline = TRUE, hidden = FALSE,
  immediate = FALSE)
}
\arguments{
\item{location}{The existing object that the new shiny object will
be placed relative to.}

\item{id}{the id of the new shiny object.}

\item{label}{the displayed label.}

\item{placement}{Object placement relative to location object.}

\item{choices}{the available choices.}

\item{selected}{the selected choice at initialization.}

\item{choiceNames}{displayed names of the choices.}

\item{choiceValues}{the server side value of the choices.}

\item{inline}{should the choices be inline or vertical.}

\item{hidden}{Should the object be hidded when created. Default = FALSE.}

\item{immediate}{Should the object be created immediately. Default = FALSE.}
}
\description{
Inserts a new shiny checkbox group object from the
server code. Wraps the insert UI function.
}
\author{
Nathan Vaughan
}
\keyword{UI}
\keyword{insertion}
