% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ui.R
\name{newColumn}
\alias{newColumn}
\title{new column}
\usage{
newColumn(location, id, placement = "beforeEnd", width = 4,
  hidden = FALSE, immediate = TRUE)
}
\arguments{
\item{location}{The existing object that the new shiny object will
be placed relative to.}

\item{id}{the id of the new shiny object.}

\item{placement}{Object placement relative to location object.}

\item{width}{column width an integer between 1-12.}

\item{hidden}{Should the object be hidded when created. Default = FALSE.}

\item{immediate}{Should the object be created immediately. Default = FALSE.}
}
\description{
Inserts a new shiny column object from the
server code. Wraps the insert UI function.
}
\author{
Nathan Vaughan
}
\keyword{UI}
\keyword{insertion}
