% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ui.R
\name{newFluidRow}
\alias{newFluidRow}
\title{new row}
\usage{
newFluidRow(location, id, placement = "beforeEnd", hidden = FALSE,
  immediate = TRUE)
}
\arguments{
\item{location}{The existing object that the new shiny object will
be placed relative to.}

\item{id}{the id of the new shiny object.}

\item{placement}{Object placement relative to location object.}

\item{hidden}{Should the object be hidded when created. Default = FALSE.}

\item{immediate}{Should the object be created immediately. Default = FALSE.}
}
\description{
Inserts a new shiny row object from the
server code. Wraps the insert UI function.
}
\author{
Nathan Vaughan
}
\keyword{UI}
\keyword{insertion}
