% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ui.R
\name{newPlot}
\alias{newPlot}
\title{new plot}
\usage{
newPlot(location, id, x = 1:10, y = rnorm(10), height = "400px",
  title = "Main", xlab = "X", ylab = "Y", placement = "beforeEnd",
  input = input, output = output, session = session,
  hidden = FALSE, immediate = FALSE)
}
\arguments{
\item{location}{The existing object that the new shiny object will
be placed relative to.}

\item{id}{the id of the new shiny object.}

\item{x}{vector of x values}

\item{y}{vector of y values}

\item{height}{height of the plot}

\item{title}{plot title text}

\item{xlab}{x axis subtitle text}

\item{ylab}{y axis subtitle text}

\item{placement}{Object placement relative to location object.}

\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}

\item{hidden}{Should the object be hidded when created. Default = FALSE.}

\item{immediate}{Should the object be created immediately. Default = FALSE.}
}
\description{
Inserts a new shiny plot object from the
server code. Wraps the insert UI function.
}
\author{
Nathan Vaughan
}
\keyword{UI}
\keyword{insertion}
