% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ui.R
\name{newSlider}
\alias{newSlider}
\title{new slider input}
\usage{
newSlider(location, id, titleText, placement = "beforeEnd", min = 0,
  max = 1, value = 0, step = 0.01, sep = ",", hidden = FALSE,
  immediate = FALSE)
}
\arguments{
\item{location}{The existing object that the new shiny object will
be placed relative to.}

\item{id}{the id of the new shiny object.}

\item{titleText}{the displayed label.}

\item{placement}{Object placement relative to location object.}

\item{min}{minimum allowed value.}

\item{max}{maximum allowed value.}

\item{value}{the displayed value.}

\item{step}{step size for value choices.}

\item{sep}{the separator for thousands between display values. default = ",".}

\item{hidden}{Should the object be hidded when created. Default = FALSE.}

\item{immediate}{Should the object be created immediately. Default = FALSE.}
}
\description{
Inserts a new shiny slider input object from the
server code. Wraps the insert UI function.
}
\author{
Nathan Vaughan
}
\keyword{UI}
\keyword{insertion}
