% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ui.R
\name{newTextAreaInput}
\alias{newTextAreaInput}
\title{new Text input area}
\usage{
newTextAreaInput(location, id, titleText, placement = "beforeEnd",
  cols = 12, rows = 20, hidden = FALSE, immediate = FALSE)
}
\arguments{
\item{location}{The existing object that the new shiny object will
be placed relative to.}

\item{id}{the id of the new shiny object.}

\item{titleText}{the displayed label.}

\item{placement}{Object placement relative to location object.}

\item{cols}{a number of columns}

\item{rows}{a number of rows}

\item{hidden}{Should the object be hidded when created. Default = FALSE.}

\item{immediate}{Should the object be created immediately. Default = FALSE.}
}
\description{
Inserts a new shiny text input area object from the
server code. Wraps the insert UI function.
}
\author{
Nathan Vaughan
}
\keyword{UI}
\keyword{insertion}
