% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ui.R
\name{newTextSlider}
\alias{newTextSlider}
\title{new text slider}
\usage{
newTextSlider(location, id, label, placement = "beforeEnd", choices,
  selected, hidden = FALSE, immediate = FALSE)
}
\arguments{
\item{location}{The existing object that the new shiny object will
be placed relative to.}

\item{id}{the id of the new shiny object.}

\item{label}{the displayed label.}

\item{placement}{Object placement relative to location object.}

\item{choices}{available options.}

\item{selected}{selected option.}

\item{hidden}{Should the object be hidded when created. Default = FALSE.}

\item{immediate}{Should the object be created immediately. Default = FALSE.}
}
\description{
Inserts a new shiny text slider input object from the
server code. Wraps the insert UI function.
}
\author{
Nathan Vaughan
}
\keyword{UI}
\keyword{insertion}
