% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_ui.R
\name{newValueInput}
\alias{newValueInput}
\title{new value input}
\usage{
newValueInput(location, id, value = NULL, placement = "beforeEnd",
  label = NULL, min = 0, max = 1, step = 0.001, hidden = FALSE,
  immediate = FALSE)
}
\arguments{
\item{location}{The existing object that the new shiny object will
be placed relative to.}

\item{id}{the id of the new shiny object.}

\item{value}{the displayed value.}

\item{placement}{Object placement relative to location object.}

\item{label}{the displayed label.}

\item{min}{minimum allowed value.}

\item{max}{maximum allowed value.}

\item{step}{step size for value choices.}

\item{hidden}{Should the object be hidded when created. Default = FALSE.}

\item{immediate}{Should the object be created immediately. Default = FALSE.}
}
\description{
Inserts a new shiny value input object from the
server code. Wraps the insert UI function.
}
\author{
Nathan Vaughan
}
\keyword{UI}
\keyword{insertion}
