% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peeling.R
\name{peeling}
\alias{peeling}
\title{The peeling algorithm}
\usage{
peeling(hgm, vars_nanes_nb_size, vars_def, hg_rel_names, elim_order)
}
\arguments{
\item{hgm}{The incidence matrix of the hypergraph (bipartite graph) describing the relations between the nodes of the graph. The names of the rows are are the names of the variables (nodes). Each column describes a relation between the variables by a (0,1) vector. A "1" indicates that a variable belongs to the relation and a "0" not.}

\item{vars_nanes_nb_size}{a table of 3 columns containing the name, number and size of all variables.}

\item{vars_def}{A list of the variables and their values obtaine by concatening the infovaluenames of all the relations.}

\item{hg_rel_names}{The names of the relations, which are objects of class "bcaspec".}

\item{elim_order}{The order of elimination of the variables. A vector of length nrows(hgm). The last one is the variable of interest.}

\item{showgraph}{= FALSE: Don't show graph.}
}
\value{
A bca class object.
}
\description{
An implementation of the peeling algorithm of Augustine Kong.\cr
}
\details{
The peeling algorithm...
}
\examples{
complete
peeling(x)
}
\references{
\itemize{
\item Almond, R. G. (1989) Fusion and Propagation of Graphical Belief Models: An Implementation and an Example. Ph. D. Thesis, the Department of Statistics, Harvard University. 288 pages (algorithm's description in pages 52-53).
}
}
\author{
Claude Boivin, Stat.ASSQ
}
