% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plautrans.R
\name{plautrans}
\alias{plautrans}
\title{Plausibility transformation applied on the distribution of singletons}
\usage{
plautrans(x)
}
\arguments{
\item{x}{A belief function in its bca form or the normalized result of the combination of two or more belief functions (see \code{\link{nzdsr}}.}
}
\value{
The matrix of singletons with the plausibility transformation added in last column.
}
\description{
The plausibility transformation is used to obtain the probability distribution associated to a belief function.
}
\details{
First, we compute the belief and plausibility measures on all the singletons of the frame of discernment. The probability distribution of the singletons is derived from the plausibility measures of the singletons.
}
\examples{
 
x <- bca(f=matrix(c(0,1,1,1,1,0,1,1,1),nrow=3, byrow = TRUE), m=c(0.2,0.5, 0.3), cnames =c("a", "b", "c"), n=1)
plautrans(x)
}
\references{
Cobb, B. R. and Shenoy, P.P. (2006). On the plausibility transformation method for translating belief function models to probability models. Journal of Approximate Reasoning, 41(3), April 2006, 314--330.
}
\author{
Claude Boivin, Stat.ASSQ
}
