% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_assessment.R
\name{prepareAssessment}
\alias{prepareAssessment}
\title{Prepare assessment}
\usage{
prepareAssessment(input, output, session, output.prep, starter.prep,
  data.prep, control.prep, forecast.prep, pars.prep, disp.Desc, dir.prep)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}

\item{output.prep}{the output file list object of the assesment being preped.}

\item{starter.prep}{the starter file list object of the assesment being preped.}

\item{data.prep}{the data file list object of the assesment being preped.}

\item{control.prep}{the control file list object of the assesment being preped.}

\item{forecast.prep}{the forecast file list object of the assesment being preped.}

\item{pars.prep}{the parameter file list object of the assesment being preped.}

\item{disp.Desc}{the display file list object of the assesment being preped.}

\item{dir.prep}{the directory of the assesment being preped.}
}
\description{
This takes raw assessment files and prepares them for
use in the shiny decision support tool by including
needed parameters such as retention and discard
parameters even if they were not included in the original
assessment.
}
\author{
Nathan Vaughan
}
\keyword{assessment}
\keyword{projection}
