% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDSM.R
\name{printDSM}
\alias{printDSM}
\alias{bcaPrint}
\title{Simple printing of the \code{tt} matrix and mass values of a Dempster-Shafer Model}
\usage{
printDSM(x, remove = FALSE)

bcaPrint(x, remove = FALSE)
}
\arguments{
\item{x}{A list of class bcaspec.}

\item{remove}{Default = FALSE. Put = TRUE to exclude subsets with zero mass.}
}
\value{
A table of subsets with their associated mass. Subsets are identified by row names.
}
\description{
This utility function does a simple printing of a DSM  (basic chance assignment).
}
\examples{
z <- DSM(tt = matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, 
byrow = TRUE), m = c(0.2,0.5, 0.3), 
cnames = c("a", "b", "c"), idvar = 1)
printDSM(z)
y <- DSM(tt=matrix(c(1,0,0,1,1,1),nrow = 2, 
byrow = TRUE), m = c(0.6,0.4), include_all = TRUE,
cnames = c("a", "b", "c"),varnames = "y", idvar = 1)
printDSM(y)
printDSM(y, remove = TRUE)
x <- DSM(tt=matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, 
byrow = TRUE), m = c(0.2,0.5, 0.3),
cnames = c("a", "b", "c"), idvar = 1, method = "ezt-m")
printDSM(x)
}
\author{
Claude Boivin
}
