% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ss_files.R
\name{rd_ctl}
\alias{rd_ctl}
\title{read control file}
\usage{
rd_ctl(file, data_input)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{data_input}{List object created by \code{\link{rd_data}}.}
}
\description{
read Stock Synthesis control file into list object in R
}
\seealso{
\code{\link{rd_starter}}, \code{\link{rd_forecast}},
\code{\link{rd_data}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Nathan Vaughan
}
\keyword{data}
\keyword{import}
