% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ss_files.R
\name{rd_data}
\alias{rd_data}
\title{read data file}
\usage{
rd_data(file, verbose = FALSE, echoall = FALSE, section = NULL)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{verbose}{Should there be verbose output while running the file?
Default=FALSE.}

\item{echoall}{Debugging tool (not fully implemented) of echoing blocks of
data as it is being read.}

\item{section}{Which data set to read. Only applies for a data.ss_new file
created by Stock Synthesis. Allows the choice of either expected values
(section=2) or bootstrap data (section=3+). Leaving default of section=NULL
will read input data, (equivalent to section=1).}
}
\description{
read Stock Synthesis data file into list object in R
}
\seealso{
\code{\link{rd_starter}}, \code{\link{rd_forecast}},
\code{\link{rd_ctl}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Ian Taylor
}
\keyword{data}
\keyword{import}
