% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ss_files.R
\name{rd_display}
\alias{rd_display}
\title{read display details file}
\usage{
rd_display(file, forecast.internal = forecast.orig,
  data.internal = data.orig)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{forecast.internal}{List object created by \code{\link{rd_forecast}}.}

\item{data.internal}{List object created by \code{\link{rd_data}}.}
}
\description{
read display details file into list object in R
}
\seealso{
\code{\link{rd_starter}}, \code{\link{rd_ctl}}, \code{\link{rd_forecast}},
\code{\link{rd_data}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Nathan Vaughan
}
\keyword{data}
\keyword{import}
