% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ss_files.R
\name{rd_forecast}
\alias{rd_forecast}
\title{read forecast file}
\usage{
rd_forecast(file = "forecast.ss", Nfleets, Nareas, Nseas,
  verbose = FALSE)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{Nfleets}{Number of fleets in the assessment.}

\item{Nareas}{Number of areas in the assessment.}

\item{Nseas}{Number of seasons in the assessment.}

\item{verbose}{Should there be verbose output while running the file?}
}
\description{
read Stock Synthesis forecast file into list object in R
}
\seealso{
\code{\link{rd_starter}}, \code{\link{rd_ctl}},
\code{\link{rd_data}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Ian Taylor
}
\keyword{data}
\keyword{import}
