% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ss_files.R
\name{rd_par}
\alias{rd_par}
\title{read parameter file}
\usage{
rd_par(file)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}
}
\description{
read Stock Synthesis parameter file into list object in R
}
\seealso{
\code{\link{rd_starter}}, \code{\link{rd_ctl}}, \code{\link{rd_forecast}},
\code{\link{rd_data}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Nathan Vaughan
}
\keyword{data}
\keyword{import}
