% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ss_files.R
\name{rd_starter}
\alias{rd_starter}
\title{read starter file}
\usage{
rd_starter(file = "starter.ss", verbose = FALSE)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{verbose}{Should there be verbose output while running the file?}
}
\description{
read Stock Synthesis starter file into list object in R
}
\seealso{
\code{\link{rd_forecast}}, \code{\link{rd_data}},
\code{\link{rd_ctl}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Ian Taylor
}
\keyword{data}
\keyword{import}
