% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ss_files.R
\name{ReadCatch}
\alias{ReadCatch}
\title{Read catch values}
\usage{
ReadCatch(output, yrs)
}
\arguments{
\item{output}{SS ouput list created by rd_output.}

\item{yrs}{the years of data to be returned.}
}
\description{
Return a timeseries of fleet catches and important
population status metrics.
}
\author{
Nathan Vaughan
}
\keyword{auxillary}
\keyword{functions}
