% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_interface.R
\name{readManagementData}
\alias{readManagementData}
\title{read management data}
\usage{
readManagementData(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
read management data values from assessment to populate
the default values in the shiny interface
these values identify the future allocation, selectivity,
and management target values for the fishery.
}
\author{
Nathan Vaughan
}
\keyword{building}
\keyword{interface}
