% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_interface.R
\name{resetForeVals}
\alias{resetForeVals}
\title{Reset forecast parameter values}
\usage{
resetForeVals(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Resets the forecast values of the shiny interface
to their default values
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
