% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_interface.R
\name{resetTabElements}
\alias{resetTabElements}
\title{Reset tab values}
\usage{
resetTabElements(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Resets the tab element values of the shiny interface
to their default values
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
