% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment.R
\name{RunAppliedForecast}
\alias{RunAppliedForecast}
\title{Run applied forecast}
\usage{
RunAppliedForecast(input, output, session, kobeComp = 1)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}

\item{kobeComp}{a designator to the saved KOBE results
file name. Default=1}
}
\description{
Runs stock synthesis with -nohess and no parameter
estimation in order to produce projections. This
process is run only once with either catch or F
at a fixed value which may be user specified or
a ratio of the target forecast calculated value.
}
\author{
Nathan Vaughan
}
\keyword{assessment}
\keyword{projection}
