% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_interface.R
\name{runSelectAssessment}
\alias{runSelectAssessment}
\title{Select assessment}
\usage{
runSelectAssessment(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Imports and organizes data for selected assessment
controls building the interface and setting default
values from original assessment.
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
