% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment.R
\name{RunTargetForecast}
\alias{RunTargetForecast}
\title{Run target forecast}
\usage{
RunTargetForecast(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Runs stock synthesis with -nohess and no parameter
estimation in order to produce projections. This
process is looped until the desiged management
targets are obtained (i.e. SPR=0.3 or MSY)
}
\author{
Nathan Vaughan
}
\keyword{assessment}
\keyword{projection}
