% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assessment.R
\name{runUpdateAssessment2}
\alias{runUpdateAssessment2}
\title{Update assessment}
\usage{
runUpdateAssessment2(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Runs an updated assessment using all the
user specified input adjustments to projection
parameters.
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
