% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_outputs.R
\name{saveCurrResults}
\alias{saveCurrResults}
\title{Save results}
\usage{
saveCurrResults(input, output, session, baseSaveName)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}

\item{baseSaveName}{the .}
}
\description{
This saves the output figures and data to a
user specified destination of their machine.
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
