% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_interface.R
\name{ScaleAllocations}
\alias{ScaleAllocations}
\title{Scale Allocations}
\usage{
ScaleAllocations(input, output, session)
}
\arguments{
\item{input}{shiny input values list.}

\item{output}{shiny output values list.}

\item{session}{the current shiny session.}
}
\description{
Scales all allocation elements so that
appropriate values sum to one.
}
\author{
Nathan Vaughan
}
\keyword{interface}
\keyword{management}
