% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnames.R
\name{ssnames}
\alias{ssnames}
\title{Construct subsets names from column names of a bca.}
\usage{
ssnames(x)
}
\arguments{
\item{x}{A bca specification}
}
\value{
subsets_names A list of names.
}
\description{
Construct subsets names from column names of a bca.
}
\examples{
y1 <- bca(tt = matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, 
byrow = TRUE), m = c(0.2,0.5, 0.3), 
cnames = c("a", "b", "c"),  
varnames = "x", idvar = 1) 
ssnames(y1)
}
\author{
Claude Boivin
}
