% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSR.R
\name{SSR}
\alias{SSR}
\title{One dimensional State Space representation}
\usage{
SSR(varnames = NULL, idvar = NULL, size = NULL, cnames = NULL)
}
\arguments{
\item{varnames}{The name of the variable. Generated if omitted.}

\item{idvar}{The number given to the state space. A number is necessary to manage relations between variables of multidimensional state spaces and enable computations on a graph. 0 if omitted.}

\item{size}{The number of elements of the state space representation \eqn{\Theta}.}

\item{cnames}{A character vector containing the names (labels) of the elements of the state space \eqn{\Theta}. The length must be equal to the \code{size} parameter. If omitted, names are generated.}
}
\value{
y A list called a SSR for "State Space Representation": \itemize{
  \item infovar  The number of the variable and the size of the state space.
  \item varnames  The name of the variable.
  \item valuenames  A list of length 1 consisting of the name of the variable with the names (labels) of the elements of the state space.
  }
}
\description{
Function \code{SSR} is used to specify a set \eqn{\Theta} of possible values.\cr
The set \eqn{\Theta} is called the state space.
}
\examples{
ss1 <- SSR("x1", 1,4,c("a","b","c","d"))

}
\references{
\itemize{
\item dDmpster (2007)...
\item Gonng (2024)...
}
}
\author{
Claude Boivin
}
