% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superBca.R
\name{superBca}
\alias{superBca}
\title{simple support bca}
\usage{
superBca(x, y, a, y0 = 0, flip = TRUE)
}
\arguments{
\item{x}{}

\item{y}{}

\item{a}{}

\item{y0}{}

\item{flip}{}
}
\value{
z a list with three elements \itemize{
 \item tt 
 \item qq 
 \item z
}
}
\description{
simple support bca
}
\details{
simple support bca
}
\examples{
1
}
\author{
Peiyuan Zhu
}
