% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{superBcaFastPlauSingleton}
\alias{superBcaFastPlauSingleton}
\title{superBcaFastPlauSingleton is a C++ algorithm aimed to optimize the computation of multiple support functions defined on very large frames of discernment}
\description{
superBcaFastPlauSingleton is a C++ algorithm aimed to optimize the computation of multiple support functions defined on very large frames of discernment
}
