% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superBcaPlauSingleton.R
\name{superBcaPlauSingleton}
\alias{superBcaPlauSingleton}
\title{Fast combination of multiple simple support functions}
\usage{
superBcaPlauSingleton(x, y, a, y0 = 0, flip = TRUE, normalize = TRUE)
}
\arguments{
\item{x}{A binary matrix of simple support functions define on a same frame of discernment (Fod).}

\item{y}{A column matrix \code{y} of TRUE/FALSE values.}

\item{a}{The mass value allotted to each simple support function. All support functions have the same mass.}

\item{y0}{A value used to check which rows of the matrix \code{x} are to be inverted. Set at 0 (FALSE).}

\item{flip}{Parameter used when some rows of the table \code{x} need to be reversed. The default value is TRUE (check rows).}

\item{normalize}{Parameter used when some rows of the vector \code{pl} need to be normalized. The default value is TRUE (check rows).}
}
\value{
pl Plausibility vector
}
\description{
Fast combination of multiple simple support functions
}
\details{
This function handles the case of combining several simple support functions defined on a very large frame of discernment (Fod). The simple support functions are arranged row by row in a binary array \code{x}. The number of rows of the array is equal to the number of support functions to be combined.The support functions can be defined as is or by their complement when it's faster to code. A column matrix \code{y} of TRUE/FALSE values is used to identify the rows of the array that need to be inverted.
}
\examples{
1
}
\author{
Peiyuan Zhu
}
