% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superset.R
\name{superset}
\alias{superset}
\title{Find the q value of the smallest superset of a vector}
\usage{
superset(node, w)
}
\arguments{
\item{node}{TO BE DOCUUMENTED}

\item{w}{TO BE DOCUUMENTED}
}
\description{
Find the q value of the smallest superset of a vector
}
