% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{treeFast}
\alias{treeFast}
\title{TreeFast is an algorithm aimed to optimize the computation of multiple support functions defined on very large frames of discernment}
\description{
TreeFast is an algorithm aimed to optimize the computation of multiple support functions defined on very large frames of discernment
}
