% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateDSM.R
\name{truncateDSM}
\alias{truncateDSM}
\alias{bcaTrunc}
\title{Truncation of the mass function of a DSM}
\usage{
truncateDSM(x, seuil, use_ssnames = FALSE)

bcaTrunc(x, seuil, use_ssnames = FALSE)
}
\arguments{
\item{x}{A DSM to truncate.}

\item{seuil}{A treshold value}

\item{use_ssnames}{Put TRUE to use ssnames parameteer instead of description matrix. Default = FALSE.}
}
\value{
tr_x The DSM object truncated.
}
\description{
When working with large frames of discernment, the DSM resulting of repeated application of Dempster's Rule of Combination can become quite large. One way to handle this situation is to group together subsets whose mass is less than a small threshold value. The function \code{DSMtrunc} is used for this purpose.\cr
}
\examples{
x <- DSM(tt = matrix(c(0,1,0,0, 
0,0,1,1,
1,1,0,0,
1,0,1,0,
0,1,1,0,
1,1,1,1),ncol=4, byrow = TRUE), m = c(0.2, 0.5, 0.06, 0.04, 0.03, 0.17),
cnames = c("a", "b", "c", "d"))
printDSM(x)
tr_x <- truncateDSM(x, seuil = 0.1)
printDSM(tr_x)

}
\author{
Claude Boivin
}
