% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttHelper.R
\name{ttHelper}
\alias{ttHelper}
\title{helper to the construction of the *tt* matrix of a *bca*}
\usage{
ttHelper(x, card, cnames, nvar = 1, binary = FALSE, formula = NULL)
}
\arguments{
\item{x}{A list... each element , the positions of 1 (TRUE) columns from 1 (left) to number of elements (right)}

\item{card}{For the single variable case. The number of elements of the frame of discernment. Maximum 1023.}

\item{cnames}{For the single variable case. A character vector containing the names of the elements of the frame of discernment \eqn{\Theta}. The length must be equal to the number of elements of \eqn{\Theta}. The names are first searched in the \code{valuenames} parameter. If NULL, column names of the matrix \code{tt}.}

\item{nvar}{number of variables. Default = 1.}

\item{binary}{Logical parameter. TRUE if multiple binary variables. Default = FALSE.}

\item{formula}{A valid propositional formula. For the multiple binary variables only.}
}
\value{
A tt matrix.
}
\description{
The tt matrix is a (0,1) or boolean matrix. As the frame of discernment increases, it becomes more and more difficult to define the subsets by sequences of (0,1) - (true, false). Function *ttHelper* tries to facilitate this task.
}
\examples{
# Single variable case
ym2 <- list(c(1,5,6), c(2,3,5,6), c(4), c(3,5), c(1,2,5,4), c(1:6) )
clabels <- c("a", "b", "c", "d", "e", "f")
ym2_enc <- ttHelper(ym2, card = 6, cnames = clabels)
rownames(ym2_enc) <- nameRows(ym2_enc)
ym2_enc
# multiple binary variables examples
union <- ttHelper(nvar = 2, binary = TRUE, formula = c("Peter | Mary"))
rownames(union) <- nameRows(union)
union
# Implication example
ifRainNoRoadworks <- ttHelper(nvar = 2, binary = TRUE, formula = c("!Rain | RdW") )
rownames(ifRainNoRoadworks) <- nameRows(ifRainNoRoadworks)
ifRainNoRoadworks

}
\author{
Claude Boivin
}
