% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmatrix.R
\name{ttmatrix}
\alias{ttmatrix}
\title{Construct a description matrix from a list of subsets names.}
\usage{
ttmatrix(x)
}
\arguments{
\item{x}{A list of names}
}
\value{
ttmat A corresponding logical description matrix
}
\description{
Construct a description matrix from a list of subsets names.
}
\examples{
subsets_names <- list(c("b", "c"), "b", c("a", "b", "c"))
ttmatrix(subsets_names)
}
\author{
Claude Boivin
}
