% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmatrixFromMarginal.R
\name{ttmatrixFromMarginal}
\alias{ttmatrixFromMarginal}
\title{Construct tt matrix of a bca from marginal probabilities}
\usage{
ttmatrixFromMarginal(
  marg_probs,
  min_prob = 0,
  max_prob = 2,
  from_above = FALSE,
  simple = FALSE
)
}
\arguments{
\item{marg_probs}{marginal probabilities}

\item{from_above=TRUE}{whether to cut marginal probabilities from above}

\item{simple=TRUE}{whether to use simple support function}
}
\value{
matrix of possible subsets obtained from the cuts
}
\description{
Construct tt matrix of a bca from marginal probabilities
}
\examples{
x <- c(2,2,1.5,1.2,1,0,0)
ttmatrixFromMarginal(x, FALSE)
}
\author{
Peiyuan Zhu
}
