% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmatrixFromTT.R
\name{ttmatrixFromTT}
\alias{ttmatrixFromTT}
\title{Construct a tt matrix of focal elements from a list of tt matrices}
\usage{
ttmatrixFromTT(TT, valuenames)
}
\arguments{
\item{TT}{List of tt matrices}

\item{valuenames}{Vector of valuenames}
}
\value{
tt A corresponding logical description matrix of the focal elements
}
\description{
Construct a tt matrix of focal elements from a list of tt matrices
}
\examples{
TT<-list(matrix(c(0,1,1,1,1,0,1,1,1),nrow = 3, byrow = TRUE),
 tt = matrix(c(1,0,0,1,1,1),nrow = 2, byrow = TRUE))
ttmatrixFromTT(TT)
}
\author{
Peiyuan Zhu
}
