% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ss_files.R
\name{wrt_ctl}
\alias{wrt_ctl}
\title{write control file}
\usage{
wrt_ctl(file, data_input, control_input, overwrite = TRUE)
}
\arguments{
\item{file}{Filename either with full path or relative to working directory.}

\item{data_input}{List object created by \code{\link{rd_data}}.}

\item{control_input}{List object created by \code{\link{rd_ctl}}.}

\item{overwrite}{Should existing files be overwritten? Default=TRUE.}
}
\description{
write Stock Synthesis control file into list object in R
}
\seealso{
\code{\link{rd_starter}}, \code{\link{rd_forecast}},
\code{\link{rd_data}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Nathan Vaughan
}
\keyword{data}
\keyword{export}
