% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ss_files.R
\name{wrt_data}
\alias{wrt_data}
\title{write data file}
\usage{
wrt_data(datlist, outfile, overwrite = TRUE, verbose = FALSE)
}
\arguments{
\item{datlist}{List object created by \code{\link{rd_data}}.}

\item{outfile}{Filename for where to write new data file.}

\item{overwrite}{Should existing files be overwritten? Default=TRUE.}

\item{verbose}{Should there be verbose output while running the file?}
}
\description{
write Stock Synthesis data file from list object in R which was probably
created using \code{\link{rd_data}}
}
\seealso{
\code{\link{rd_starter}},
\code{\link{rd_forecast}}, \code{\link{rd_ctl}},
\code{\link{wrt_starter}}, \code{\link{wrt_forecast}},
\code{\link{wrt_data}}, \code{\link{wrt_ctl}}
}
\author{
Ian Taylor
}
\keyword{data}
\keyword{export}
