% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ss_files.R
\name{wrt_display}
\alias{wrt_display}
\title{write display file}
\usage{
wrt_display(mylist, file = paste(getwd(), "Display_Descriptions.txt", sep
  = "/"), overwrite = TRUE, verbose = FALSE)
}
\arguments{
\item{mylist}{List object created by \code{\link{rd_display}}.}

\item{file}{Filename and directory for new forecast file.
Default="Display_Descriptions.txt".}

\item{overwrite}{Should existing files be overwritten? Default=TRUE.}

\item{verbose}{Should there be verbose output while running the file?
Default=FALSE}
}
\description{
write display descriptions file into list object in R
}
\seealso{
\code{\link{rd_starter}}, \code{\link{rd_forecast}},
\code{\link{rd_data}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Ian Taylor
}
\keyword{data}
\keyword{export}
