% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ss_files.R
\name{wrt_forecast}
\alias{wrt_forecast}
\title{write forecast file}
\usage{
wrt_forecast(mylist, dir = NULL, file = "forecast.ss",
  overwrite = TRUE, verbose = FALSE)
}
\arguments{
\item{mylist}{List object created by \code{\link{rd_forecast}}.}

\item{dir}{Directory for new forecast file. Default=NULL (working directory).}

\item{file}{Filename for new forecast file. Default="forecast.ss".}

\item{overwrite}{Should existing files be overwritten? Default=TRUE.}

\item{verbose}{Should there be verbose output while running the file?
Default=FALSE}
}
\description{
write Stock Synthesis forecast file into list object in R
}
\seealso{
\code{\link{rd_starter}}, \code{\link{rd_forecast}},
\code{\link{rd_data}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Ian Taylor
}
\keyword{data}
\keyword{export}
