% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ss_files.R
\name{wrt_par}
\alias{wrt_par}
\title{write parameter file}
\usage{
wrt_par(file, pars_input, overwrite = TRUE)
}
\arguments{
\item{file}{Filename for new parameter file.}

\item{pars_input}{List object created by \code{\link{rd_par}}.}

\item{overwrite}{Should existing files be overwritten? Default=TRUE.}
}
\description{
write Stock Synthesis parameter file into list object in R
}
\seealso{
\code{\link{rd_starter}}, \code{\link{rd_forecast}},
\code{\link{rd_data}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Nathan Vaughan
}
\keyword{data}
\keyword{export}
