% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ss_files.R
\name{wrt_starter}
\alias{wrt_starter}
\title{write starter file}
\usage{
wrt_starter(mylist, dir = NULL, file = "starter.ss",
  overwrite = TRUE, verbose = FALSE, warn = FALSE)
}
\arguments{
\item{mylist}{List object created by \code{\link{rd_starter}}.}

\item{dir}{Directory for new starter file. Default=NULL (working directory).}

\item{file}{Filename for new starter file. Default="starter.ss".}

\item{overwrite}{Should existing files be overwritten? Default=TRUE.}

\item{verbose}{Should there be verbose output while running the file?
Default=FALSE.}

\item{warn}{Print warning if overwriting file?}
}
\description{
write Stock Synthesis starter file from list object in R which was probably
created using \code{\link{rd_starter}}
}
\seealso{
\code{\link{rd_starter}}, \code{\link{rd_forecast}},
\code{\link{rd_ctl}}, \code{\link{wrt_starter}},
\code{\link{wrt_forecast}}, \code{\link{wrt_data}},
\code{\link{wrt_ctl}}
}
\author{
Ian Taylor
}
\keyword{data}
\keyword{export}
