% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToDSM.R
\name{addToDSM}
\alias{addToDSM}
\alias{addTobca}
\title{Add some elements of 0 mass to an existing DSM.}
\usage{
addToDSM(x, tt)

addTobca(x, tt)
}
\arguments{
\item{x}{A DSM (see \code{\link{DSM}}).}

\item{tt}{A matrix constructed in a boolean style (0,1) or a boolean matrix. The number of columns of the matrix \code{tt} must match the number of columns of the \code{tt} matrix of \code{x} (see \code{\link{DSM}}). Each row of the matrix identify a subset of the state space.}
}
\value{
x The original DSM \code{x} augmented with the added subsets defined by \code{tt}.
}
\description{
You may want to add to a previously defined DSM some elements or some subsets of the state space (the set of possible values). The mass of these additional elements or subsets may be set to zero. This feature is useful when one wants to examine the plausibility measure of elements or subsets of zero mass.
}
\examples{
 
y <- DSM(tt = matrix(c(1,0,0,1,1,1),nrow=2, byrow = TRUE), 
m = c(0.6, 0.4),  cnames = c("a", "b", "c"), idvar = 1)
addToDSM(y, matrix(c(0,1,0,0,0,1, 0,1,1), nrow = 3, byrow = TRUE))
x <- DSM(tt = matrix(c(0,1,1,1,1,0,1,1,1),nrow=3, 
byrow = TRUE), m=c(0.2,0.5, 0.3), 
cnames = c("a", "b", "c"), idvar = 1)
xy <- dsrwon(x,y)
xy1 <- addToDSM(nzdsr(xy), matrix(c(0,1,0,0,0,1), nrow = 2, byrow = TRUE))
xy1
# add all singletons to a DSM
addToDSM(x, tt = diag(rep(1, ncol(x$tt) ) )  ) 

}
\author{
Claude Boivin
}
