% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ads}
\alias{ads}
\title{The Captain's Problem. ads: Relation between variables Arrival (A), Departure delay (D) and Sailing delay (S)}
\format{An integer matrix with 18 rows and 17 columns.
\describe{
  \item{[1,c(1,2)]}{value = 0, not used}
  \item{[1,3:17]}{Identification numbers of the three variables. Column 3 to 9: variable 1; colunn 10 to 13: variable 2; colunn 14 to 17: variable 3.}
  \item{nospec}{identification number of the specification}
  \item{m}{the value of the spécification, a number between 0 and 1}
  \item{6}{1 if 6 is part of the specification, 0 otherwise}
  \item{5}{1 if 5 is part of the specification, 0 otherwise}
  \item{4}{1 if 4 is part of the specification, 0 otherwise}
  \item{3}{1 if 3 is part of the specification, 0 otherwise}
  \item{2}{1 if 2 is part of the specification, 0 otherwise}
  \item{1}{1 if 1 is part of the specification, 0 otherwise}
  \item{0}{1 if 0 is part of the specification, 0 otherwise}
  }}
\source{
\url{https://www.researchgate.net/publication/265993533_Fusion_and_Propagation_in_Graphical_Belief_Models}
}
\usage{
ads
}
\description{
This dataset is the tt matrix establishing the relation A = D + S, where A = {0:6}, D = {0:3} and S = {0:3}. All the elements (a,d,s) of (A x D x S) where a = d + s is true form a subset with a mass value of 1. To construct the tt matrix, we put the variables A, D, S side by side, as in a truth table representation.
}
\author{
Claude Boivin, Stat.ASSQ
}
\keyword{datasets}
