% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{bound}
\alias{bound}
\title{Find inf, sup of a lattice}
\usage{
bound(x, method = NULL)
}
\arguments{
\item{method}{Return supremum or infimum}

\item{tt}{Binary matrix encoding the lattice from which the supremum or infimum is found}
}
\value{
b Binary vector of sup or inf
}
\description{
Find inf, sup of a lattice
}
\author{
Peiyuan Zhu
}
