% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{captain_result}
\alias{captain_result}
\title{The Captain's Problem. \code{swr}: Result of the evaluation of the Hypergraph at node Arrival (A)}
\format{
A list of 8 elements, of class bcaspec.
}
\source{
{Almond, R.G. [1988] Fusion and Propagation in Graphical Belief Models. Computing Science and Statistics: Proceedings of the 20th Symposium on the Interface. Wegman, Edward J., Gantz, Donald T. and Miller, John J. (ed.). American Statistical Association, Alexandria, Virginia. pp 365--370.}
}
\usage{
captain_result
}
\description{
This dataset is the \code{tt} bca resulting from the combination of the relations of the hypergraph and marginalization at node Arrival (A).
}
\author{
Claude Boivin, Stat.ASSQ
}
\keyword{datasets}
