% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{closure}
\alias{closure}
\title{Augment list of binary vectors with closure elements}
\arguments{
\item{ttxl}{A list of binary vectors}

\item{computeJoin}{= true: to compute join closure. Default = TRUE}
}
\value{
A list of binary vectors including the closure elements
}
\description{
Augment list of binary vectors with closure elements
}
