% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{closure}
\alias{closure}
\title{Augment list of binary vectors with closure elements}
\arguments{
\item{ttxl}{list of binary vectors}
}
\description{
Augment list of binary vectors with closure elements
}
