% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{closure}
\alias{closure}
\title{Augment a sparse binary matrix with closure elements}
\arguments{
\item{ttx}{A binary matrix}

\item{computeJoin}{= true: to compute join closure. Default = TRUE}

\item{display_progress}{= true: to compute join closure. Default = FALSE}
}
\value{
A binary matrix including the closure elements
}
\description{
Augment a sparse binary matrix with closure elements
}
\examples{
ttx <- matrix(c(0,1,1,1,1,0,1,1,1), nrow=3, byrow = TRUE)
tty <- closure(ttx, computeJoin = FALSE)
}
