% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNode.R
\name{createNode}
\alias{createNode}
\title{Change depth to appropriate calculation}
\usage{
createNode(x, q, index = NULL)
}
\arguments{
\item{x}{A boolean or binary vector coerced to bit vector.}

\item{q}{A commonnality value associated to \code{x}.}

\item{index}{TO BE DOCUMMENTED.}
}
\value{
A list of length 5 containing the following components:  \itemize{
\item x The binary table.
\item q The commonalities.
\item left The left side of the tree.
\item right The right side of the tree.
\item depth The depth of the tree..
}
}
\description{
Change depth to appropriate calculation
}
\details{
Depth equals the greatest index corresponding to 1 in the bit vector \code{x}.
}
\examples{
# TO DO

}
\author{
Peiyuan Zhu
}
