% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsrwon.R
\name{dsrwon}
\alias{dsrwon}
\title{Combination of two belief functions}
\usage{
dsrwon(x, y, infovarnames = NULL, relnb = NULL)
}
\arguments{
\item{x}{A belief function in its bca form (see \code{\link{bca}}).}

\item{y}{A belief function represented by its bca form.}

\item{infovarname}{A name can be given to the resulting variable. Named "v1" if missing.
@param relnb A number can be given to the resulting relation. if omitted, the numbers of the input relations will be kept.}
}
\value{
A list of five elements: \itemize{
  \item $combination: The matrix f of focal elements with a column of masses added.
  \item $I12 A table of intersections between subsets.
  \item $sort_order Indices for the sort of the propositions.
  \item $con: the measure of conflict between beliefs.
  \item $n: the node number.
 \item $inforel. A two column matrix containing variable numbers and the depth of the relation.
  }
  @details The relation number of the x input is given to the result.
}
\description{
The unnormalized Dempster's rule is used to combine two belief functions Bel1 and Bel2 defined on the same frame of discernment and represented by their respective basic chance assignments \code{x} and \code{y}. Dempster's rule of combination is applied. The normalization is not done, leaving the choice to the user to normalize the results or not (See \code{\link{nzdsr}}).
}
\details{
A function (inters) does the intersection of the propositions. Another function (doubles) remove duplicate elements from the result of the intersection of the M propositions of the first bca with the N propositions of the second bca. The two bcas \code{a} and \code{b} must be defined on the same frame of discernment for the intersection of propositions to be meaningful.
}
\examples{
x1 <- bca(f=matrix(c(0,1,1,1,1,0,1,1,1),nrow=3, byrow = TRUE), m=c(0.2,0.5, 0.3), cnames =c("a", "b", "c"),  infovarnames = "x", n=1)
x2 <- bca(f=matrix(c(1,0,0,1,1,1),nrow=2, byrow = TRUE), m=c(0.6, 0.4),  cnames = c("a", "b", "c"),  infovarnames = "x", n=1)
dsrwon(x1,x2)
}
\references{
Shafer, G., (1976). A Mathematical Theory of Evidence. Princeton University Press, Princeton, New Jersey, p. 57-61: Dempster's rule of combination.
}
\author{
Claude Boivin, Stat.ASSQ
}
