% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elim.R
\name{elim}
\alias{elim}
\title{Reduction of a relation}
\usage{
elim(rel, x)
}
\arguments{
\item{rel}{the relation to modify}

\item{x}{the number of the variable to eliminate}
}
\value{
The reducted relation
}
\description{
This function eliminates a chosen variable from a relation and calculate the bcaspec of the remaining variables.
}
\author{
Claude Boivin, Stat.ASSQ
}
