% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initsing.R
\name{initsing}
\alias{initsing}
\title{Initialization of a frame of discernment}
\usage{
initsing(size, cnames, infovarnames = NULL)
}
\arguments{
\item{size}{The number of elements of the frame of discernment.}

\item{cnames}{A character vector of the names of the elements of the frame.}

\item{infovarnames}{(optional) A name for the variable can be supplied.}
}
\value{
An object of class "bcaspec".
}
\description{
The purpose of this function is to initialize the basic chance assignment of all the singletons of a frame at a starting value of 0. Each singleton has a zero mass. The matrix of singletons is completed by the frame which receive a mass of one. One use of this function is to apply the plausibility transformation to a belief function and obtain its corresponding probability distribution (See \code{\link{plautrans}}).
}
\examples{
initsing(3,cnames = c("a","b","c"))
initsing(3,cnames = c("a","b","c"), infovarnames = "choices")
}
\author{
Claude Boivin, Stat.ASSQ
}
