% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertNode.R
\name{insertNode}
\alias{insertNode}
\title{Insert bit vector and value into a node}
\usage{
insertNode(node1, node2)
}
\arguments{
\item{node1}{TO BE DOCUUMENTED}

\item{node2}{An existing tree structure}
}
\value{
The tree augmented structure.
}
\description{
node1$x is a boolean or binary vector coerced to bit vector. 
node1$q is a commonnality value associated to \code{x}.
}
\details{
# TO DO
}
\examples{
# TO DO
}
\author{
Peiyuan Zhu
}
