% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrayToMatrix.R
\name{marrayToMatrix}
\alias{marrayToMatrix}
\title{Transformation of an array data to a matrix-represented relation}
\usage{
marrayToMatrix(mtt, infovar)
}
\arguments{
\item{mtt}{The matrix tt of the relation in  multiarray format}

\item{infovar}{specification of the number and size of each variable}
}
\value{
tt The matrix representation of the data.
}
\description{
The array representation or product space representation is converted to the matrix representation of the corresponding relation.
}
\examples{
wr_tt <- matrix(c(0,1,rep(0,5),rep(c(1,0),2),1,1,0,1,0,
rep(1,3),0,1,0,rep(1,6)), ncol=4, byrow = TRUE)
colnames(wr_tt) <- c("rWdy Ry", "rWdy Rn", "rWdn Ry", "rWdn Rn")
 wr_spec = matrix(c(1:7, 0.0476, 0.7619, 0.1905, 0,0,0,0), 
 ncol = 2, dimnames = list(NULL, c("specnb", "mass"))) 
 wr_infovar = matrix(c(4,5,2,2), ncol = 2, 
 dimnames = list(NULL, c("varnb", "size")) )
 wr_rel <- list(tt=wr_tt, con=0.16, spec=wr_spec,
  infovar=wr_infovar, 
  infovaluenames= list( RdWorks=c("rWdy", "rWdn") , Rain=c("Ry", "Rn")))
 class(wr_rel)="bcaspec"
 mtt <- array(c(0,1,0,0,0,0,0,1,0,1,0,1,1,0,1,0,1,1,1,0,1,0,1,1,1,1,1,1), c(2,2,7), dimnames = list( RdWorks=c("rWdy", "rWdn") , Rain=c("Ry", "Rn"), ev=1:7))
}
\author{
Claude Boivin, Stat.ASSQ
}
