% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrayToMatrix.R
\name{marrayToMatrix}
\alias{marrayToMatrix}
\title{Transformation of an array data to its matrix representation}
\usage{
marrayToMatrix(mtt)
}
\arguments{
\item{mtt}{The matrix tt of the relation in  multiarray format}
}
\value{
tt The matrix representation of the data.
}
\description{
The array representation or product space representation is converted to the matrix representation of the corresponding relation.
}
\examples{
 mtt <- array(c(1,0,0,0,1,1,0,0,1,0,0,1,1,0,1,0,1,1,0,1,1,1,1,0,1,0,1,1,1,1,1,1), c(2,2,8), 
 dimnames = list( RdWorks=c("Wy", "Wn") , Rain=c("Ry", "Rn"), ev=1:8) )
 print(z <- marrayToMatrix(mtt))

}
\author{
Claude Boivin, Stat.ASSQ
}
