% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFromMarginal.R
\name{mFromMarginal}
\alias{mFromMarginal}
\title{Construct m vector of a bca from marginal probabilities}
\usage{
mFromMarginal(marg_probs, a = 1e-05, min_prob = 0, max_prob = 2, simple = TRUE)
}
\arguments{
\item{marg_probs}{vector of marginal probabilities}

\item{a=1e-5}{probability that the sample is reliable}

\item{min_prob=0}{lower bound on marginal probabilities}

\item{max_prob=2}{upper bound on marginal probabilities}

\item{simple=TRUE}{whether to use simple support function}
}
\value{
vector of probability masses obtained from uniformly sampling the cut
}
\description{
Construct m vector of a bca from marginal probabilities
}
\examples{
p <- c(2,2,1.5,1.2,1,0,0)
mFromMarginal(p, FALSE)
}
\author{
Peiyuan Zhu
}
