% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFromQQ.R
\name{mFromQQ}
\alias{mFromQQ}
\title{Construct a mass vector from qq function.}
\usage{
mFromQQ(qq, tt)
}
\arguments{
\item{qq}{Commonality function}

\item{tt}{logical description matrix from ttmatrixFromQQ}
}
\value{
m A corresponding mass vector
}
\description{
Construct a mass vector from qq function.
}
\examples{
tt<- t(matrix(c(1,0,1,1),ncol = 2))
m<- c(.9,.1)
cnames <- c("yes","no")
x<- bca(tt, m, cnames=cnames)
mFromQQ(x$qq, x$tt)
}
\author{
Peiyuan Zhu
}
