% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFromQQRecursive.R
\name{mFromQQRecursive}
\alias{mFromQQRecursive}
\title{Construct a mass vector from qq function and ttmatrix of focal elements recursively.}
\usage{
mFromQQRecursive(qq, tt)
}
\arguments{
\item{qq}{Commonality function}

\item{tt}{logical description matrix of focal elements from ttmatrixFromTT}
}
\value{
m A corresponding mass vector
}
\description{
Construct a mass vector from qq function and ttmatrix of focal elements recursively.
}
\examples{
tt<- t(matrix(c(1,0,1,1),ncol = 2))
m<- c(.9,.1)
cnames <- c("yes","no")
x<- bca(tt, m, cnames=cnames)
mFromQQ(x$qq, x$tt)
}
\author{
Peiyuan Zhu
}
