% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameRows.R
\name{nameRows}
\alias{nameRows}
\title{Naming the rows of a matrix of focal elements}
\usage{
nameRows(f)
}
\arguments{
\item{f}{A boolean matrix or a matrix constructed in a boolean style (0,1). The names of the columns are the elements of a frame of discernment \eqn{\Theta}.}
}
\value{
The result is a character vector of the names of the focal elements of the matrix \code{f}. The length of the result is \code{nrows(f)}. The set of all elements is called "frame", to avoid to long a string. The empty set is named "u00f8". The "+" sign is used instead of the "|" to represent the logical "or" operation.
}
\description{
This function uses the column names  of a matrix \code{f} of focal elements to construct the names of the rows.
}
\examples{
f <- matrix(c(0,0,0,1,0,0,0,0,1,1,0,1,1,1,1),ncol=3, byrow = TRUE)
colnames(f) <- c("A","B","C")
rownames(f) <-nameRows(f)
f
f2 <- matrix(c(0,0,0,1,0,0,0,0,1,1,0,1),ncol=3, byrow = TRUE)
colnames(f2) <- c("A2","B2","C2")
rownames(f2) <-nameRows(f2) 
f2
}
\author{
Claude Boivin, Stat.ASSQ
}
