% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productSpace.R
\name{productSpace}
\alias{productSpace}
\title{Representation of belief functions in their product space}
\usage{
productSpace(zswr)
}
\arguments{
\item{zswr}{A list of the bcaspec class}
}
\description{
When two or more belief functions are related, we need to form a product space representation.
}
\examples{
 ttfw= matrix(c(1,0,1,0,0,1,0,1,1,1,1,1),nrow=3, byrow = TRUE, dimnames =list(NULL, c("foul", "fair", "foul", "fair")) )
 specfw = matrix(c(1,1,2,0.8,0.8,0.2), ncol = 2, dimnames = list(NULL, c("specnb", "mass"))) 
 infovarfw =matrix(c(5,7,2,2), ncol = 2, dimnames = list(NULL, c("varnb", "size")) )
 xyz <- list(tt=ttfw, spec=specfw, infovar=infovarfw)
(productSpace(xyz))
}
\author{
Claude Boivin, Stat.ASSQ
}
