% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reduction}
\alias{reduction}
\alias{aplRDV}
\alias{\code{reduction}}
\alias{returns...}
\title{Obtain the summary of a vector for any operator.}
\usage{
reduction(x, f = "+")
}
\arguments{
\item{x}{The data...}

\item{f}{The operator.}
}
\description{
Obtain the summary of a vector for any operator.
}
\references{
\itemize{
  \item G. Helzer. (1989): \emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
 \item  L. Gilman and A. J. Rose.(1974): \emph{APL an Interactive Approach}, Second Edition, John Wiley, New-York.
 }
 @examples 
 @export
}
\seealso{
Jan de Leeuw and Masanao Yajima: \url{https://rpubs.com/deleeuw/158476}.
}
\author{
Original author of the aptRDV function: Jan de Leeuw 
      \url{http://www.codecollector.net/view/8A8D9395-0F66-4706-A23E-C588151E8423-95744-0000429BCF33A153}.
}
