% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shape}
\alias{shape}
\alias{aplShape}
\title{Obtain dimensions of an array or length of a vector with a single command}
\usage{
shape(a)
}
\arguments{
\item{a}{An array or a vector.}
}
\value{
The dimension of the array \code{a} or the length of the vector \code{a}.
}
\description{
\code{shape} returns the dimension of given array or returns the length of a given vector. The function is taken from the project APL in R (\url{https://rpubs.com/deleeuw/158476}).
}
\examples{
shape(array(c(1:6), c(2,3)))
shape(c("a", "b"))
}
\references{
\itemize{
\item Jan de Leeuw and Masanao Yajima (March 07, 2016) \emph{APL in R (Version 009)}, Source code. \url{https://rpubs.com/deleeuw/158476}
  \item G. Helzer. (1989): \emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
 \item  L. Gilman and A. J. Rose.(1974): \emph{APL an Interactive Approach}, Second Edition, John Wiley, New-York.
 }
}
\author{
Claude Boivin
}
