% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superBca.R
\name{superBca}
\alias{superBca}
\title{Fast combination of many support functions}
\usage{
superBca(x, y, a, y0 = 0, flip = TRUE, tree_type = "single")
}
\arguments{
\item{x}{The binary matrix of support functions}

\item{y}{A logical column matrix which row of the matrix x must be inverted}

\item{a}{The mass value of the simple support functions}

\item{y0}{The value used to check for the rows of x to invert.}

\item{flip}{TRUE by default. Function will check for the presence of rows to invert.}
}
\value{
z a list with three elements \itemize{
 \item tt 
 \item qq 
 \item z
}
}
\description{
Fast combination of many support functions
}
\details{
The case of many support functions with the same mass.
}
\examples{
1
}
\author{
Peiyuan Zhu
}
