% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabresul.R
\name{tabresul}
\alias{tabresul}
\title{Prepare a table of results}
\usage{
tabresul(x, singletonsOnly = FALSE, removeZeroes = FALSE)
}
\arguments{
\item{x}{A belief function in its bca form, generally the normalized result of the combination of two or more belief functions (see \code{\link{nzdsr}}.}

\item{singletonsOnly}{= TRUE reduces the table of results to elementary events.}

\item{removeZeroes}{= TRUE removes focal elements with 0 mass.}
}
\value{
A list of three elements: \itemize{
  \item $mbp: The table of focal elements with the addition of the their associated mass, degree of belief, plausibility and the plausibility ratio
  \item $con The measure of conflict between focal elements.
  \item $n The node number.
  }
}
\description{
This utility function gather the bca representation of a belief function and the measures of belief and plausibility in one table.
}
\examples{
 
x <- bca(f=matrix(c(0,1,1,1,1,0,1,1,1),nrow=3, byrow = TRUE), m=c(0.2,0.5, 0.3), cnames =c("a", "b", "c"), n=1)
y <- bca(f=matrix(c(1,0,0,1,1,1),nrow=2, byrow = TRUE), m=c(0.6, 0.4),  cnames = c("a", "b", "c"), n=1)
tabresul(nzdsr(xy)) 
print("Remove focal elements with 0 mass")
tabresul(nzdsr(xy), removeZeroes = TRUE)
tabresul(nzdsr(xy), singletonsOnly = TRUE)
print("Add the two missing singletons")
xy1 <- addTobca(nzdsr(xy), matrix(c(0,1,0,0,0,1), nrow=2, byrow = TRUE))
tabresul(xy1, singletonsOnly = TRUE)
}
\author{
Claude Boivin, Stat.ASSQ
}
