% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmatrix.R
\name{ttmatrix}
\alias{ttmatrix}
\title{Construct a description matrix from a list of subsets names.}
\usage{
ttmatrix(x, sparse = "no")
}
\arguments{
\item{x}{A list of names}

\item{sparse=c("yes", "no")}{whether to use sparse matrix}
}
\value{
ttmat A corresponding logical description matrix
}
\description{
Construct a description matrix from a list of subsets names.
}
\examples{
subsets_names <- list(c("b", "c"), "b", c("a", "b", "c"))
ttmatrix(subsets_names)
znames <- list("empty", "a", c("b", "c"), c("a", "b"), c("a", "b", "c") )
print(ttmatrix(znames) )
print(ttmatrix(znames, sparse = "yes") )
}
\author{
Claude Boivin
}
