% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmatrixFromMarginal.R
\name{ttmatrixFromMarginal}
\alias{ttmatrixFromMarginal}
\title{Construct tt matrix of a bca from marginal probabilities}
\usage{
ttmatrixFromMarginal(marg_probs, from_above = TRUE)
}
\arguments{
\item{marg_probs}{marginal probabilities}

\item{from_above}{whether to cut marginal probabilities from above}
}
\value{
matrix of possible subsets obtained from the cuts
}
\description{
Construct tt matrix of a bca from marginal probabilities
}
\examples{
p <- c(2,2,1.5,1.2,1,0,0)
ttmatrixFromMarginal(p, FALSE)
}
\author{
Peiyuan Zhu
}
