% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmatrixFromQQ.R
\name{ttmatrixFromQQ}
\alias{ttmatrixFromQQ}
\title{Construct a description matrix from a qq vector}
\usage{
ttmatrixFromQQ(qq, n, cnames, sparse = "no")
}
\arguments{
\item{qq}{Commonality function}

\item{n}{Dimension of the frame}

\item{cnames}{A character vector containing the names of the elements of the frame of discernment}

\item{sparse}{= c("yes","no") whether to use sparse matrix. Default = "no".}
}
\value{
tt A corresponding logical description matrix
}
\description{
This utility function is called when we need to reconstruct a tt matrix, e.g. to print a DSM.
}
\examples{
tt<- t(matrix(c(1,0,1,1),ncol = 2))
m<- c(.9,.1)
cnames <- c("yes","no")
x<- bca(tt, m, cnames=cnames, method="ezt-m")
ttmatrixFromQQ(x$qq,as.integer(x$infovar[1,2]), cnames)
}
\author{
Peiyuan Zhu
}
