% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmatrixFromQQ.R
\name{ttmatrixFromQQ}
\alias{ttmatrixFromQQ}
\title{Construct a description matrix from qq function.}
\usage{
ttmatrixFromQQ(qq, n, valuenames)
}
\arguments{
\item{qq}{Commonality function}

\item{n}{Dimension of the frame}

\item{valuenames}{Vector of valuenames}
}
\value{
ttmat A corresponding logical description matrix
}
\description{
Construct a description matrix from qq function.
}
\examples{
tt<- t(matrix(c(1,0,1,1),ncol = 2))
m<- c(.9,.1)
cnames <- c("yes","no")
x<- bca(tt, m, cnames=cnames)
ttmatrixFromQQ(x$qq,as.integer(x$infovar[1,2]), unlist(x$valuenames))
}
\author{
Peiyuan Zhu
}
